services/leonardo/pyproject.toml
[tool.poetry]
name = "leonardo"
version = "0.1.0"
description = ""
authors = ["Nir Mantzur <nirooma@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
django-extensions = "^3.1.3"
ipython = "^7.28.0"
psycopg2-binary = "^2.9.1"
channels = "^3.0.4"
channels-redis = "^3.3.1"
django-storages = "^1.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
